Almost functional shadow implementation.
Shadows basically work as long as the light is looking down the z plane
and the model is on the ground. Need to fix the code which finds near
and far planes as it breaks when the camera is rotated out of alignment
with z axis, and need to fix the camera's view, as from its pov it
renders the scene to only the top of the screen, cutting off even the
top of the model if its y increases.