Fix renderpass/swapchain write after write
This seems buggy and weird. Added a comment regarding this in main.c.
The semaphore wait stage between color commands and presentation are not
behaving as expected. Using ALL_COMMANDS is required in place of
COLOR_ATTACHMENT_OUTPUT in order to prevent the sync hazard. Using the
logical or of all stages apparently constituting ALL_COMMANDS gives
errors regarding stages being handled which are not present in the
pipeline.